+++ CHANGELOG.md
@@ -7,6 +7,7 @@ ChangeLog
 - Added Django 5 compatibility
 - README fix
 - form_valid hook added on login view
+- One user logout functionality added
 
 
 1.2 2017-12-18